+++ cucumber-scala/src/test/scala/io/cucumber/scala/steps/errors/incorrectobjecthooks/IncorrectObjectHooksDefinition.scala
@@ -8,11 +8,17 @@ object IncorrectObjectHooksDefinition extends ScalaDsl {
   // On a single line to avoid difference between Scala versions for the location
 
   // A body that does not return Unit => interpreted as missing body
+  BeforeAll { 22 }
+  
+  // A body that does not return Unit => interpreted as missing body
   Before { 1 }
 
   // A body that does not return Unit => interpreted as missing body
   BeforeStep { "toto" }
 
+  // A body that does not return Unit => interpreted as missing body
+  AfterAll { 66 }
+  
   // A body that does not return Unit => interpreted as missing body
   After { 33 }